+++ Command/DumpConfigurationCommand.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace steevanb\PhpUrlTest\Command;
 
-use steevanb\PhpUrlTest\Configuration\Dumper\ConsoleConfigurationDumper;
+use steevanb\PhpUrlTest\{
+    Configuration\Dumper\ConsoleConfigurationDumper,
+    CreateUrlTestServiceTrait
+};
 use Symfony\Component\Console\{
     Command\Command,
     Input\InputArgument,
@@ -23,6 +26,12 @@ class DumpConfigurationCommand extends Command
 
         $this
             ->setName('urltest:configuration:dump')
+            ->addOption(
+                'configuration',
+                'conf',
+                InputOption::VALUE_OPTIONAL,
+                'Configuration file name'
+            )
             ->addOption('recursive', 'r', InputOption::VALUE_OPTIONAL, 'Set recursive if path is a directory.', 'true')
             ->addArgument('path', InputArgument::REQUIRED, 'Configuration file name, or directories separated by ",".')
             ->addArgument('ids', InputArgument::OPTIONAL, 'UrlTest identifiers preg pattern to test.');
@@ -31,15 +40,17 @@ class DumpConfigurationCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $ids = $input->getArgument('ids') === null ? null : explode(',', $input->getArgument('ids'));
-        $service = $this->createUrlTestService(
+        $urlTestService = $this->createUrlTestService(
             $input->getArgument('path'),
-            $input->getOption('recursive') === 'true'
+            $input->getOption('recursive') === 'true',
+            $input->getOption('configuration')
         );
 
-        $dumper = (new ConsoleConfigurationDumper($output))->dumpGlobal($service, $ids);
-        foreach ($service->getTests($ids ?? []) as $urlTest) {
+        $dumper = (new ConsoleConfigurationDumper($output))->dumpConfiguration($urlTestService, $ids);
+
+        foreach ($urlTestService->getTests($ids ?? []) as $urlTest) {
             $output->writeln('');
-            $dumper->dump($urlTest, $output->getVerbosity() >= OutputInterface::VERBOSITY_VERBOSE);
+            $dumper->dumpUrlTest($urlTest);
         }
 
         return 0;